Add const for identifying main Connect project
This was introduce because some APIs, such as the
GetConnectJobs method, allow for retrieval across
all projects by not including the project query param.
Since the Connect <Main> project is retrieved by
an empty string project query param there needed to be
a way to differentiate between the <Main> project and
not specifying a project at all so the project query param can
be removed altogether.